Lots of updates
- Updated to iOS 8
- Made it look prettier.  Added in tabbar icons
- Added app icon
- Fixed margins for pictures, description and business information
- changed retail replacement value to appraisal value
- Please note, I tried adding the ability to turn the phone sideways,
but it really messes up the spacing and a lot of the controls when
turned to landscape.  This is why I kept it in portrait.

One last thing left to do is ability to see previous appraisals.  This
takes a bit of time, so that will be added at a later date.